+++ .gitignore
@@ -1,2 +1,2 @@
 node_modules/
-build/
+build/node_modules/
